JKRAram: back ARAM with host memory + real DMA copy (2D archive load works)
The native ARAM seam (ar_dsp_impl.cpp) only allocated addresses; ARQ DMA was a
no-op completion that moved no bytes, so ARAM was always empty. That was harmless
while the only ARAM user was audio (native_jas bypasses it), but the 2D/UI ARCHIVE
path genuinely DMAs through ARAM: TApplication::setupThreadFuncLogo streams
guide.szs/game_6.szs into ARAM (gArBkGuide/gArBkConsole), and SMSSwitch2DArchive
copies them back out (Yaz0-decompressing ARAM -> main RAM) to mount the "guide" 2D
volume. With ARAM empty, JKRFileLoader::getVolume("guide") was null and
TConsoleStr::load OSPanic'd (the stage-15/file-select setupObjects blocker).

Make ARAM real:
- ar_dsp_impl.cpp: ARInit calloc's a 16 MB host buffer; ARQPostRequest/ARStartDMA
  do a direction-aware memcpy between host main-RAM pointers and ARAM offsets,
  fail-fast (OSPanic) on an out-of-range ARAM offset.
- LP64: a main-memory address is a 64-bit host pointer that cannot round-trip
  through the SDK's u32 DMA args (it truncated to a wild pointer). Widen the whole
  ARAM DMA address chain to pointer width (new ARMemAddr = uintptr_t native / u32 on
  GC, mirroring the existing ARQRequestRef): JKRAMCommand mSrc/mDst, JKRAramPcs /
  orderSync / orderAsync / prepareCommand, ARQPostRequest source/dest, ARStartDMA
  mainmem_addr, and drop the truncating (u32)(uintptr_t) casts at every call site
  (JKRAram, JKRAramStream, JKRDvdAramRipper, JASDvdThread). On GC uintptr_t == u32 so
  the on-target layout/behavior is unchanged.

Verified: ar_dsp round-trips bytes through ARAM (render/ctest), and the live boot
mounts "guide" then game_6 (correct decompressed RARC + sizes; mountFixed ok),
clearing the TConsoleStr panic. SB_ARC_DBG dumps the ARAM header + mount result.

+++ src/JSystem/JAudio/JASystem/JASDvdThread.cpp
@@ -234,8 +234,8 @@ int Dvd::loadToAramDvdTMain(void* param)
 			call->unk28 += buffersize;
 			call->unk2C -= buffersize;
 		}
-		ARQPostRequest(&req[arq_index], 0x12345678, 0, 1, (u32)(uintptr_t)buf,
-		               (u32)(uintptr_t)call->unk24, batchSize, &aramDmaFinish);
+		ARQPostRequest(&req[arq_index], 0x12345678, 0, 1, (uintptr_t)buf,
+		               (uintptr_t)call->unk24, batchSize, &aramDmaFinish);
 		++bufferFull;
 		++arq_index;
 		arq_index %= 4;
